backend/requirements.txt
azure-cli
azure-cosmos
azure-identity
Flask
Flask-cors
openai
semantic_kernel
tiktoken
langchain
langchain-community
azure-search-document

# openai-1.14.1 pydantic-2.6.4 python-dotenv-1.0.1 referencing-0.30.2 semantic-kernel-0.9.3b1

